requirements.txt
flask>=3.0.0
pillow>=10.0.0
ultralytics>=8.0.0
python-snap7>=1.2.0
